feat: implement Jurusan model, standardize database relationships, and add polymorphic link support across evaluation models.

+++ database/migrations/2024_07_21_031803_create_targets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('targets', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('id_indikator');
-            $table->float('value');
+            $table->decimal('value', 10, 2);
             $table->timestamps();
 
             $table->foreign('id_indikator')->references('id')->on('indikators')->onDelete('cascade');
